Correcciones en la calase Tecnico

+++ ejercicios_repaso_H_P_repaso2/ejercicios_repaso_H_P_repaso2_02/src/ejercicios_repaso_h_p_repaso2_02/Tecnico.java
@@ -20,7 +20,7 @@ public class Tecnico extends Empleado {
     }
 
     //Constructor con todos los atributos
-    public Tecnico(int salario, String nombre, String apellido, int edad, double productividad) {
+    public Tecnico(double salario, String nombre, String apellido, int edad, double productividad) {
         super(salario, nombre, apellido, edad);
         this.productividad = productividad;
     }
